Refactor components into own files
